feat[GROUP_LIST]: Refactor group list component and add visibility icon; update hooks for improved state management
